requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic==1.10.14
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
httpx==0.27.0
pytest==7.4.4
email-validator==2.2.0
alembic==1.13.1
psycopg2-binary==2.9.9
